# feat: return the jsonpath of json detected fields