--- a/pkg/logproto/logproto.proto
+++ b/pkg/logproto/logproto.proto
@@ -529,6 +529,7 @@ message DetectedField {
   uint64 cardinality = 3;
   repeated string parsers = 4 [(gogoproto.jsontag) = "parsers"];
   bytes sketch = 5 [(gogoproto.jsontag) = "sketch,omitempty"];
+  repeated string jsonPath = 6 [(gogoproto.jsontag) = "jsonPath,omitempty"]; // Array representing original JSON path (e.g., ["user", "id"] for field "user_id")
 }
 
 message DetectedLabelsRequest {