… pkg/logproto/logproto.pb.go
… pkg/logql/syntax/ast_test.go
… pkg/querier/queryrange/detected_fields_test.go